backend/pyproject.toml
[project]
name = "fastapi-vue-backend"
version = "0.1.0"
description = "FastAPI backend template for FastAPI + Vue full-stack starter"
authors = [{ name = "Template Maintainers" }]
requires-python = ">=3.13"
readme = "README.md"
dependencies = [
  "fastapi>=0.115.0,<0.116",
  "uvicorn[standard]>=0.30.0,<0.31",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.3,<3",
  "sqlalchemy[asyncio]>=2.0.34,<2.1",
  "alembic>=1.14.0,<1.15",
  "psycopg[binary]>=3.2.0,<3.3",
  "passlib[bcrypt]>=1.7.4,<1.8",
  "bcrypt>=4.0,<4.2",
  "pyjwt[crypto]>=2.9.0,<3",
  "python-multipart>=0.0.9,<0.1",
  "redis>=5.0.1,<6",
  "celery[redis]>=5.3.0,<6.0",
  "httpx>=0.27.0,<0.28",
  "email-validator>=2.2.0,<3",
  "structlog>=24.1.0,<25",
  "opentelemetry-sdk>=1.30.0,<1.31",
  "opentelemetry-exporter-otlp-proto-http>=1.30.0,<1.31",
  "debugpy>=1.8.0,<1.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<8.4",
  "pytest-asyncio>=0.24.0,<0.25",
  "pytest-dotenv>=0.5.2,<0.6",
  "pytest-cov>=5.0.0,<5.1",
  "ruff>=0.7.0,<0.8",
  "mypy>=1.13.0,<1.14",
  "types-redis>=4.6.0,<4.7",
  "types-passlib>=1.7.7,<1.8",
  "types-pyjwt>=1.7.1,<1.8",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = ["fastapi-vue-backend[dev]"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-config --strict-markers"
asyncio_mode = "auto"
pythonpath = ["app"]
testpaths = ["app/tests"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "I", "UP", "B", "ARG", "C4", "SIM", "TID"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"app/tests/*" = ["S101", "ARG001"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]
warn_return_any = true
warn_unused_configs = true


